Fix bug in distribute_local_to_global
If a local matrix with all diagonal elements equal to zero is distributed to a global matrix, the l1 norm of the local matrix divided by the size of the local matrix is now added to those diagonal elements of the global matrix which correspond to a constrained dof. In case the entire local matrix is zero, 1 is added. Previously zero was added for both cases, possibly resulting in singular global matrices. Additionally, a test for the patched version of distribute_local_to_global has been added.
Fixes #7658