tutorials: add keywords
I am often searching on the tutorial overview page in the browser. It is
helpful to have additional keywords available to know where a feature is
used/introduced. Right now I am only adding this for the first few steps
(and only class/function names).

I can follow up on this later for the other tutorials if we agree this
makes sense.