Rectify the interface traction term
Change the sign in front of the pressure in the interface traction term and use the normal on the solid (as described in the tutorial documentation) instead of the normal on the fluid for the computation of the interface traction contribution to the system matrix. Previously, the pillar was bending in the wrong direction due to the wrong sign.