Consistently enforce inlining in simple tensor functions.
This is necessary because compiler heuristics get bad in complicated situation, replacing simple
calls to one-liners when inlined into a lot of overhead through function calls. A complicated
situation is, e.g., when the same operations on tensors are done by several callers, say a
function templated on the polynomial degree in matrix-free computations where the quadrature
point loop does some operations with tensors in the same way for all degrees.