Add a push_back member function to Quadrature<dim>.
Immersed quadrature rules are complicated to construct. It is often
done in several steps over different parts of a cell. Add a push_back
function to make it easier to construct a quadrature. Emphasize in the
documentation that the function should only be used as the quadrature
is being constructed.