Make TensorProductQuadrature a separate class.
Presently a boolean is used internally to keep track of whether
Quadrature<dim> is a tensor product or not. This makes Quadrature<dim>
have two different states which suggests that the tensor product
functionality would be better off as a separate class. Add a new class
TensorProductQuadrature and move the 1-dimensional base quadratures
from Quadrture<dim> to this new class. Derive the different quadrature
rules (QGauss, etc.) which are tensor products from
TensorProductQuadrature.