remove TimerOutput error spam
TimerOutput produces output on each MPI rank when an exception is
thrown. This makes looking at output with 100k MPI ranks somewhat
annoying. Fix this by only outputting on rank 0.
This assumes that also rank 0 triggers this exception, which might not
be true. In that case, we would likely deadlock anyways...