Insert particles globally.
Create and insert a number of particles into the collection of
particles.
This function takes a list of positions and creates a set of particles

at these positions, which are then distributed and added to the local

particle collection of a procesor. Note that this function uses

GridTools::distributed_compute_point_locations. Consequently, it can

require intense communications between the processors.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>