Label step-40 as a 'basic' tutorial program.
So far, only steps 1-6 were listed as 'basic', but step-40 has attained a similar
status in recent years since quite a number of new programs are now parallel based
on the techniques discussed there. I think it would make sense to reflect that in
the graph by giving it a highlighted style.