Also compile the complex versions of UMFPACK functions.
An attempt at simply changing what I though needed to be changed in
the CMakeLists.txt failed with linker errors, so I decided to go back
to the original UMFPACK Makefile and be more faithful that file's
structure by renaming some of the variables we have in the current
file. The diff is likely difficult to read, but the changes are
ultimately not rocket science.