Keep the 'dealii::' in 'using namespace dealii::'.
Some of our example code on the website (e.g., step-40) doesn't work since we
get rid of the preceding dealii and print

    using namespace ::LinearAlgebraPETSc;

when it should be

    using namespace dealii::LinearAlgebraPETSc;

instead. This change keeps the 'dealii::' only in this case, which currently
only occurs in step-40.cc, step-50.cc, step-55.cc, and local_integrators.h.