Add test of make_flux_sparsity_pattern in 1D.
For the version of make_flux_sparsity_pattern that takes face and cell
couplings, add a 1D test that fails for both types of DoFHandler. For
dealii::DoFHandler it fails because there is a bug in 1D and for the hp
case it fails because make_flux_sparsity_pattern isn't instantiated in
the 1D case.