Remove update_once/each() functions from FE_NedelecSZ.
These functions *used* to be part of the public interfaces of FE classes, but haven't
in a long time. They also aren't 'virtual' any more for this class. Furthermore,
all uses of these functions are of the form 'update_once() | update_each()', so
we may as well just call the 'requires_update_flags()' function that is the
current interface.