Use CMake USE_SOURCE_PERMISSIONS to install the header files
Without the option USE_SOURCE_PERMISSIONS, in some corner cases the permissions
of the header files can be modified and that can cause problems for the quotas
of certain clusters.

In certain clusters the administrator uses the setgid "s" permission to track
the quotas. Without the option USE_SOURCE_PERMISSIONS, CMake removes the
setgid permission and that can generate the error "Disk quota exceeded".