Test boost/extract_representative_set_02: add output variant
The actual final set of bounding boxes computed with boost depends on
the standard library being used. Funny enough the difference is only one
cell on one processor when switching from libc++ to libstc++. The output
seems to be stable with respect to standard library. I have not
investigated further, but I would bet that it boils down to an
(unstable) sort algorithm being used in the boost implementation.

For now, just provide an output variant. If this test fails again, we
might have to think about another alternative.