Rename namespace internals:: to internal::AffineConstraints::.
The AffineConstraints class and friends put things into a namespace 'internals'. That's
an uncommon name: We typically use 'internal::'. But a lot of the things in that namespace
also have rather generic names, so do as we often do in such cases: further qualify
the name to indicate what class the internals correspond to.