Get rid of TriaObject.
TriaObject was a class that was a barely disguised ArrayView, but the class
name wouldn't have told you that. It's not actually very difficult to replace
it by ArrayView, which is what this patch does.

In the process, rename the TriaObjects::get_object() function to
TriaObjects::get_bounding_object_indices() which is both a more
descriptive name of what it returns, and matches the recently
introduced set_bounding_object_indices() function.