tests/symengine/batch_optimizer: add output variants
Add output variants for the symengine/batch_optimizer that fails due to
changes in the libstdc++ library between version 9 and 10.

Let's simply add an output variant for the time being. The proper fix
will be to change these tests to not produce output depending on the
standard library or compiler that is being used.