lac/constraints_move: Fix undefined behavior
It is undefined behavior to access an object that has been moved. It is
also not possible to "move back" an object.

In this particular case with the recent changes to ThreadLocalStorage
and AffineConstraints we run into an exception:

constraints_move.debug: ../nptl/pthread_mutex_lock.c:117: __pthread_mutex_lock: Assertion `mutex->__data.__owner == 0' failed.

(The move operation destroys the mutex and accessing it afterwards
when we try to move back results in a crash.)