Make two thread_local variables local to a function.
They previously were 'thread_local', but that creates issues with
compiler bugs when used in header files.

In the situation here, the constructor of UUID needs to set some data
to random numbers. It currently does that by using a thread_local
MT19937 engine that is seeded by a std::random_device that is also
thread_local. But because the seed is only evaluated once per thread,
it doesn't actually matter whether the std::random_device is
thread_local or just a regular local variable: the object is set up
once and evaluated once.

The situation is more complicated for the MT19937 object: Right now,
it is set up once per thread. Because of the random seed, it generates
different random sequences on each thread. It would also generate
random sequences if there was only one static variable (namely, a
sub-sequence of what MT19937 generates), which is what this patch
does. Which sub-sequence this is is non-deterministic -- but it also
wasn't before because of the use of std::random_device(). To make
access to the MT19937 object thread-safe, we need a mutex to guard it.