CMake: Bugfix: set C++14 flag in CXX_FLAGS_SAVED
When configuring we make an attempt to set the flag -std=c++14 if we
fail to detect C++14 support. Unfortunately, we accidentally set this
flag in DEAL_II_CXX_FLAGS instead of DEAL_II_CXX_FLAGS_SAVED.

This implied that RESET_CMAKE_REQUIRED() did not include the -std=c++14
flag and consequently compile tests that relied on C++14 support failed.