Simplify code by using C-style arrays.
For static data tables of known size, there is really no good reason
to use std::array over C-style arrays. The latter are a lot easier to
read because the array bounds appear in the order in which one indexes
things; for example,
  double table[6][4];
makes it clear that
  table[i][j]
is valid if i<6 and j<4. But with std::array, one has to declare
things as
  std::array<std::array<double,4>,6> table;
which is much longer and not as easy to read because bounds appear in
the "wrong" order.

Using C-style arrays also alleviates our persistent problem of having
to add more and more curly braces :-)