Rename the write/update particle functions.
Give these functions a more systematic name. While also adjust to a more widely
used idiom: A function that takes a pointer doesn't return the updated pointer
location via that argument, but instead as an explicit return value. This idiom
is in fact used just a couple of lines below the places where these functions
are currently used in particle_handler.cc.