Mark some classes as 'extern template'.
This saves a small amount of space and, possibly, compile time. Here
is a comparison of the cumulative size of all object files for the
debug and release configurations without and with the patch:

without patch:
debug   2,414,548,192
release   459,866,272

with patch:
debug    2,404,864,416
release    459,686,888

One can also compare the sizes of the .so files. I did this only for
the release library:

without patch: 459,866,272
with patch:    459,686,888

So it's a rather minor improvement. One could do this to a lot more
classes than just the two in this patch, though, if we wanted to go
that route.