Add class RefSpaceFEFieldFunction
Add a new class RefSpaceFEFieldFunction, which similarly to
FEFieldFunction evaluates a function defined by a solution vector and a
DoFHandler, but transformed to reference space. Thus defining the
function

\hat{f}(\hat{x}) = \sum_{j=0}^{n-1} f_j \hat{\phi}_j(\hat{x})

where \hat{\phi}_j(\hat(x)) are the local reference space shape
functions, f_j the solution coefficients, and \hat{x} a point on the
reference cell.