Mark members in ParticleAccessor as 'private'.
This class is not meant to be derived from, and I *think* that the reason why
these members were only 'protected' and not 'private' is to somehow indicate
that these are members that 'friend' classes might access. But that is too
complicated an interpretation. Let's just make everything 'private', as
usual.