A better assert in FE_Q_Base::get_subface_interpolation_matrix
The previous assert checked that the element derived from FE_Q_Base
or FE_SimplexPoly. This isn't right for FE_Bernstein, which derives
from FE_Q_Base, but doesn't have support points. Instead of throwing
ExcInterpolationNotImplemented, using FE_Bernstein resulted in an
exception which was hard to understand. Change the assert so that it
checks if the incoming element has support points on the face and has
the right number of components.