Add a spacedim template parameter on ImmersedSurfaceQuadrature
To handle the case when we have a line integral immersed in a face, the
normals on ImmersedSurfaceQuadrature need to be vectors in spacedim.
The reason is that the conormal can not be computed without knowing the
normal to the surface. Add a spacedim template parameter and change the
normals to Tensor<1, spacedim>. Rewrite the documentation and modify
the test to a account for this.