disable threading in task_info.cc for oneAPI
TBB with oneAPI API has deprecated and removed the
<code>tbb::tasks</code> backend. With this it is no longer possible to
compile the following code that builds a directed acyclic graph (DAG) of
(thread parallel) tasks without a major porting effort. It turned out
that such a dynamic handling of dependencies and structures is not as
competitive as initially assumed. Consequently, this part of the matrix
free infrastructure has seen less attention than the rest over the last
years and is (presumably) rarely used.

So in case of detected oneAPI backend we simply disable threading in the
matrix free backend for now.