Fix a DoFRenumbering issue with p::s::T.
This can happen if some processors don't have any degrees of freedom (e.g., we
run in parallel with only one cell in the triangulation). This crashed in a
spectacular way since this test would ultimately call MPI_Allgather with
different types at the same time.

The correct workaround here is to verify that all processors are doing the same
type of renumbering.