CMake: remove -Wl,--as-needed flag setup
The issue with the current handling of the --Wl,--as-needed flag is the
fact that we set it internally but later on filter it (unconditionally)
for user projects.

This has the issue that it is currently not possible to set
--Wl,--as-needed by hand when needed.

Thus, let us simply remove the setup and the workarounds.