CMake: Make configure file naming consistent
Configure file are including and processed in alphabetical order.

However, in order to allow for feature dependencies we have to include
some configure files before others. We have solved this by numbering
some of the configure files in order to force them to be processed
before others.

This pull requests makes this numbering a bit more consistent by

 - numbering _all_ of the configure files with a default number of 50,
   and configure files that have to be processed earlier taking up a
   number between 40 and 00.

 - increasing the numbering from one digit to two digits to accomodate
   a possible future need to adjust the inclusion order a bit better.