Fix the MPI window functions with clang.
libc++ defines std::unique_ptr::reset() as

    _LIBCPP_INLINE_VISIBILITY
    void reset(nullptr_t = nullptr) _NOEXCEPT {
      pointer __tmp = __ptr_.first();
      __ptr_.first() = nullptr;
      if (__tmp)
        __ptr_.second()(__tmp);
    }

The result of this is that aligned_vector->elements.get() is nullptr at the
point at which we call the deleter - hence this deleter won't work correctly.

We can work around this by avoiding std::unique_ptr::get() and just using our
own member variables or ptr, which are guaranteed to be valid since they will
not be set to nullptr by std::unique_ptr::reset().