Remove SparsityPatternBase.
We only have one class which inherits from this (SparsityPattern) and,
since we instantiate templates for each sparsity pattern type, it isn't
possible to do something useful by inheriting from this class.

This was split off from SparsityPattern in
5213023ca165bfbd420423551c184a3e31886223 with the intent of writing some
more general storage classes based on SparsityPatternBase, but that has
never left the proof-of-concept stage.