CMake: add a compile_test_executables target
This commit adds a top-level compile_test_executables and modifies the
ADD_TEST() macro to let it depend on all executables generated by our
DEAL_II_PICKUP_TESTS() mechanism.

This is mainly intended for user projects and works around the issue
that simultaneous calls to ninja are not reentrant, meaning when
configuring a user project with ninja one is otherwise limited to run
tests in serial.