CMake: Remove _USER variables
Traditionally we have used two "independent" sets of variables to keep
track of feature configurations:

  <FEATURE>_INCLUDE_DIRS
  <FEATURE>_DEFINITIONS(|_RELEASE|_DEBUG)

and

  <FEATURE>_USER_INCLUDE_DIRS
  <FEATURE>_USER_DEFINITIONS(|_RELEASE|_DEBUG)

The former was intended for internal use only and the latter for
exporting whatever is necessary to user projects.

However, this has not been used at all in the last 10 years.

So let us remove this distinction and simply always export include dirs
and definitions to users. This simplification is required to switch to
our features to interface targets.