Remove a dependency in writing VTU files.
Previously, one of the lambda functions that write parts updated a variable outside
the lambda function that was later checked. This works when running everything
sequentially, but introduces a race condition when we want to run things
in parallel. Move the updating code out of the lambda function, and so
effectively make it 'const' in the sense that it no longer writes into
variables outside of the lambda function.