CMake: Remove pkg-config pc file generation
Let us remove the pkg-config configuration for the time being. This is
necessary as a preparatory switch for changing over to import/interface
targets. This has always been rarely used configuration with issues (for
example for Debian and Ubuntu we cannot even install these pkg-config
files due to distribution-specific QA violations).

We can revisit the question how to (properly) create pkg-config files
again after the changes to the build system are done.