CMake: Bugfix: only add to the linker line what is needed
The `INTERFACE_LINK_LIBRARIES` option of the Kokkos target might contain
much more than we actually need. So restrict the fixup to finding the
missing linker flag and adding this one.

This works around an issue with the regression tester where the link
interface of the Kokkos::kokkos target contains entries that make the
final link fail.