CMake: Bugfix: only export DEAL_II_GMSH_WITH_API if gmsh is configured
If the gmsh library is installed but the gmsh executable is missing we
currently disable gmsh support. This implies that we will not link
against the gmsh library.

Unfortunately, on first configure pass the variable `GMSH_WITH_API` is
still populated with a `TRUE` value and the `DEAL_II_GMSH_WITH_API`
variable gets set by accident and final linkage fails.

This issue is hard to spot because a second invocation of cmake will
cure the configure mistake (and the debian/ubuntu packages do not run
any autodetection).