CMake: avoid race conditions between concurrent prm tests
We need to serialize all tests that share a common executable target.
This involves tests with .threads=N. and .mpirun=N. annotation, as well
as tests with parameter files (that might share a common executable
target).

Running multiple variants of tests with the same target executable in
parallel triggers a race condition where the same (not yet existent)
target is built concurrently leading to undefined outcomes.