Use a new, consistent orientation scheme for simplices.
A problem with the previous version is that the line orientations depended on
the face: e.g., on face 0, line 0 is (0, 1), but on face 1 it is (1, 0). This
inconsistency causes lines to be oriented both ways in certain meshes. The fix
is to simply orient all lines from low vertex to high vertex.

For consistency I made the same change with triangles, even though the
orientation problem does not exist there.