Test gmsh/gmsh_api_(01|02|03|04): add output variants
It seems that there was a change in more recent versions of gmsh and the
mesh output changed. These output variants seem to be stable among a
number of architectures and linux distributions so I simply commit them.

The output is similar to the main version and seems to be a "superset".
(It looks like older gmsh versions removed unused points ?!)

If we ever start to increase the number of output variants again I
suggest to change the test to ".run_only" and not compare unstable
output.