CMake: enable -pedantic for language feature check
This is a funny one: gcc and clang silently accept the [[deprecated]]
and [[fallthrough]] attributes which are introduced in the C++17
standard - even when the compiler is configured only with C++14 support.

BUT, clang will issue a warning that the attributes are not supported if
compiled with -pedantic. Thus, we have to check for support of the
attributes with -pedantic set.