Return objects by value, rather than via a reference.
We now have move operators that apply to return objects, along with mandatory
return value optimization. There is no readon not to use this.

This patch applies this to some internal functions. I will write a separate
patch to make this the case for the public interfaces that call these
functions.