Test metis/metis_*: only run tests, do not compare output
We currently carry 5 output variants per test and I would have to add
another 2 output variants for all tests to pass on the regression
tester.

I think it is time that we acknowledge that metis output is not stable.

Instead let us simply check that the tests compile and run successfully
without throwing an error.