Test scalapack/scalapack_05: add output variant
It is a bit disconcerting that the computed recirpocal condition number
of the matrix changes. But given the fact that the manual computation
(first number) and the scalapack-internal function return vastly
differing numbers already tell me that the test setup simply isn't
robust.

Let us add an output variant for the scalapack_05 test. I get almost the
same number on three different configuration so I am hopeful that we can
get away with an output variant for the time being.